+++ pyefis/instruments/ai/__init__.py
@@ -30,20 +30,44 @@ from pyefis import common
 
 log = logging.getLogger(__name__)
 
+# TODO:
+#   Check TAS quality and revert to turn rate indication
+#   Fix quality indications
+#   Add quality flags to indicate the actual failures
+#   Remove internal assignments of database items.  Should be a normal
+#      Qt widget and these things done externally.
+#   Add configuration for bank angle tick sizes
+
 class AI(QGraphicsView):
     def __init__(self, parent=None):
         super(AI, self).__init__(parent)
         self.myparent = parent
+
+        # The following information is meant to be configurable from the screen
+        # definition file
+        self.fontSize = 30
+        # Number of degrees shown from top to bottom
+        self.pitchDegreesShown = 60
+        # Pitch tick mark configurations
+        self.minorDiv = 1   # Degrees between minor divisions
+        self.majorDiv = 5  # Degrees between major divisions
+        self.numberedDiv = 10 # Degrees between numbered divisions
+        # Line widths of the pitch tick marks
+        self.minorDivWidth = 10
+        self.majorDivWidth = 40
+        self.numberedDivWidth = 50
+        self.visiblePitchAngle = 15 # Amount of visible pitch angle marks
+        self.pitchOpacity = 0.6
+        # Standard rate turn bank angle indicators.
+        self.drawBankMarkers = True
+        self.bankAngleRadius = None # Radius of the bank angle markings
+        self.bankAngleMaximum = 25  # Largest bank angle that will be indicated
+
         self.setStyleSheet("border: 0px")
         self.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
         self.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
         self.setRenderHint(QPainter.Antialiasing)
         self.setFocusPolicy(Qt.NoFocus)
-        self.fontSize = 30
-        # Number of degrees shown from top to bottom
-        self.pitchDegreesShown = 60
-        self.overlayColor = QColor(Qt.green)
-        self.srTurnColor = QColor(Qt.yellow)
 
         pitch = fix.db.get_item("PITCH")
         pitch.valueChanged[float].connect(self.setPitchAngle)
@@ -60,26 +84,15 @@ class AI(QGraphicsView):
         self._AIOld = roll.old
         self._AIBad = roll.bad
         self._AIFail = roll.fail
-        self.fdrolldb = fix.db.get_item("FDROLL", wait=False, create=True)
-        self.fdpitchdb = fix.db.get_item("FDPITCH", wait=False, create=True)
-        self.fdondb = fix.db.get_item("FDON", wait=False, create=True)
-        self.fdondb.valueChanged[bool].connect(self.fdon)
-        self.fdtarget_widget = None
-        self.fdt = None
-        self.fix_tas = fix.db.get_item("TAS")
-
-    def fdon(self, v):
-        if v and self.isVisible():
-            if self.fdtarget_widget is None:
-                self.fdtarget_widget = FDTarget(QPointF(self.scene.width()/2, self.scene.height()/2), self.pixelsPerDeg)
-
-                self.fdtarget_widget.resize (120, 120)
-                self.fdt = self.scene.addWidget (self.fdtarget_widget)
-        else:
-            if self.fdt is not None:
-                self.scene.removeItem(self.fdt)
-                self.fdt = None
-                self.fdtarget_widget = None
+        alat = fix.db.get_item("ALAT")
+        alat.valueChanged[float].connect(self.setLateralAcceleration)
+        self._latAccel = alat.value
+        tas = fix.db.get_item("TAS")
+        tas.valueChanged[float].connect(self.setTrueAirspeed)
+        self._tas = tas.value
+        # We store all the pitch tick marks and text in a list so that
+        # we can adjust the opacity of the items.
+        self.pitchItems = []
 
     def resizeEvent(self, event):
         #Setup the scene that we use for the background of the AI
@@ -88,7 +101,9 @@ class AI(QGraphicsView):
                                self.height() * self.height())
         self.pixelsPerDeg = self.height() / self.pitchDegreesShown
         self.scene = QGraphicsScene(0, 0, sceneWidth, sceneHeight)
-
+        # Setup default values
+        if self.bankAngleRadius is None:
+            self.bankAngleRadius = self.height() / 3
         # Get a failure scene ready in case it's needed
         self.fail_scene = QGraphicsScene(0, 0, sceneWidth, sceneHeight)
         self.fail_scene.addRect(0,0, sceneWidth, sceneHeight, QPen(QColor(Qt.white)), QBrush(QColor(50,50,50)))
@@ -119,7 +134,7 @@ class AI(QGraphicsView):
         self.setScene(self.scene)
         gradientBrown = QLinearGradient(0, sceneHeight / 2, 0, sceneHeight)
         gradientBrown.setColorAt(0.0, QColor(105, 46, 1))
-        gradientBrown.setColorAt(0.2, QColor(244, 164, 96))
+        gradientBrown.setColorAt(0.3, QColor(244, 164, 96))
         gradientBrown.setSpread(0)
         self.brown_pen = QPen(QColor(160, 82, 45))  # Brown Color
         self.gbrown_brush = QBrush(gradientBrown)
@@ -135,26 +150,6 @@ class AI(QGraphicsView):
         self.land_rect = self.scene.addRect(0, sceneHeight / 2 + 1, sceneWidth, sceneHeight,
                            self.brown_pen, brush)
 
-        """ Not sure if this is needed or not:
-        self.bad_text = self.scene.addSimpleText("BAD", font)
-        warn_pen = QPen(QColor(255, 150, 0))
-        warn_brush = QBrush(QColor(255, 150, 0))
-        self.bad_text.setPen (warn_pen)
-        self.bad_text.setBrush (warn_brush)
-        r = self.bad_text.boundingRect()
-        self.bad_text.setPos ((sceneWidth-r.width())/2, (sceneHeight-r.height())/2)
-        if not self._AIBad:
-            self.bad_text.hide()
-
-        self.old_text = self.scene.addSimpleText("OLD", font)
-        self.old_text.setPen (warn_pen)
-        self.old_text.setBrush (warn_brush)
-        r = self.old_text.boundingRect()
-        self.old_text.setPos ((sceneWidth-r.width())/2, (sceneHeight-r.height())/2)
-        if not self._AIOld:
-            self.old_text.hide()
-            """
-
         self.setScene(self.scene)
 
         #Draw the main horizontal line
@@ -163,78 +158,101 @@ class AI(QGraphicsView):
         self.scene.addLine(0, sceneHeight / 2, sceneWidth, sceneHeight / 2, pen)
         #draw the degree hash marks
         pen.setWidth(2)
-        pen.setColor(self.overlayColor)
+        pen.setColor(Qt.white)
         w = self.scene.width()
         h = self.scene.height()
         f = QFont()
         f.setPixelSize(self.fontSize)
 
-        # This loop is for drawing the tick marks
-        for i in range(1, 10):
-            left = w / 2 - self.width() / 8
-            right = w / 2 + self.width() / 8
-            inset = self.width() / 16
+        # Draw the tick mark lines on the scene
+        for i in range(-90, 91):
+            if i == 0:  # Don't draw a line at zero
+                pass
+            elif i % self.numberedDiv == 0:
+                left = w / 2 - self.numberedDivWidth / 2
+                right = w / 2 + self.numberedDivWidth / 2
+                y = h / 2 - self.pixelsPerDeg * i
 
-            # Draw the ticks above the line
-            y = h / 2 - (self.pixelsPerDeg * 10) * i
-            self.scene.addLine(left, y, right, y, pen).setZValue(1)
-            yy = y + self.pixelsPerDeg * 5
-            self.scene.addLine(left + inset, yy, right - inset, yy, pen).setZValue(1)
-            # Draw the text for each of these
-            t = self.scene.addText(str(i * 10))
-            t.setFont(f)
-            self.scene.setFont(f)
-            t.setDefaultTextColor(self.overlayColor)
-            t.setX(right + 5)
-            t.setY(y - t.boundingRect().height() / 2)
-            t.setZValue(1)
+                l = self.scene.addLine(left, y, right, y, pen)
+                l.setZValue(1)
+                self.pitchItems.append((i, l))
+                t = self.scene.addText(str(i))
+                t.setFont(f)
+                self.scene.setFont(f)
+                t.setDefaultTextColor(Qt.white)
+                t.setX(right + 5)
+                t.setY(y - t.boundingRect().height() / 2)
+                t.setZValue(1)
+                self.pitchItems.append((i, t))
 
-            t = self.scene.addText(str(i * 10))
-            t.setFont(f)
-            self.scene.setFont(f)
-            t.setDefaultTextColor(self.overlayColor)
-            t.setX(left - (t.boundingRect().width() + 5))
-            t.setY(y - t.boundingRect().height() / 2)
-            t.setZValue(1)
+                t = self.scene.addText(str(i))
+                t.setFont(f)
+                self.scene.setFont(f)
+                t.setDefaultTextColor(Qt.white)
+                t.setX(left - (t.boundingRect().width() + 5))
+                t.setY(y - t.boundingRect().height() / 2)
+                t.setZValue(1)
+                self.pitchItems.append((i, t))
 
-            # Draw the tick marks below the line
-            y = h / 2 + (self.pixelsPerDeg * 10) * i
-            self.scene.addLine(left, y, right, y, pen).setZValue(1)
-            yy = y - self.pixelsPerDeg * 5
-            self.scene.addLine(left + inset, yy, right - inset, yy, pen).setZValue(1)
-            # Draw the text for these
-            y = h / 2 + (self.pixelsPerDeg * 10) * i
-            t = self.scene.addText(str(i * - 10))
-            t.setFont(f)
-            self.scene.setFont(f)
-            t.setDefaultTextColor(self.overlayColor)
-            t.setX(right + 5)
-            t.setY(y - t.boundingRect().height() / 2)
-            t.setZValue(1)
+            elif i % self.majorDiv == 0:
+                left = w / 2 - self.majorDivWidth / 2
+                right = w / 2 + self.majorDivWidth / 2
+                y = h / 2 - self.pixelsPerDeg * i
+                l = self.scene.addLine(left, y, right, y, pen)
+                l.setZValue(1)
+                self.pitchItems.append((i, l))
 
-            t = self.scene.addText(str(i * - 10))
-            t.setFont(f)
-            self.scene.setFont(f)
-            t.setDefaultTextColor(self.overlayColor)
-            t.setX(left - (t.boundingRect().width() + 5))
-            t.setY(y - t.boundingRect().height() / 2)
-            t.setZValue(1)
-        self.update_period = self.myparent.get_config_item('update_period')
-        if self.update_period is None:
-            self.update_period = .1
-        self.last_update_time = 0
-        self.show_standard_turn = self.myparent.get_config_item('show_standard_turn')
-        if self.show_standard_turn is None: self.show_standard_turn = True
+            elif i % self.minorDiv == 0:
+                left = w / 2 - self.minorDivWidth / 2
+                right = w / 2 + self.minorDivWidth / 2
+                y = h / 2 - self.pixelsPerDeg * i
+                l = self.scene.addLine(left, y, right, y, pen)
+                l.setZValue(1)
+                self.pitchItems.append((i, l))
+        self.setPitchItems()
+
+        # Draws the static overlay stuff to a pixmap
+        self.map = QPixmap(self.width(), self.height())
+        self.map.fill(Qt.transparent)
+        p = QPainter(self.map)
+        p.setRenderHint(QPainter.Antialiasing)
+        # change width and height to the view instead of the scene
+        w = self.width()
+        h = self.height()
+        r = self.bankAngleRadius
+        p.setPen(QColor(Qt.black))
+        p.setBrush(QColor(Qt.yellow))
+        p.drawRect(w / 4, h / 2 - 3, w / 6, 6)
+        p.drawRect(w - w / 4 - w / 6, h / 2 - 3, w / 6, 6)
+        p.drawRect(w / 2 - 3, h / 2 - 3, 8, 8)
+
+        p.setPen(QColor(Qt.black))
+        p.setBrush(QColor(Qt.white))
+
+        p.translate(w / 2, h / 2)
+        triangle = QPolygon([QPoint(7, - r + 25),
+                             QPoint(-7, -r + 25),
+                             QPoint(0, - (r - 10))])
+        p.drawPolygon(triangle)
+
+
+        self.overlay = self.map.toImage()
+
         self.redraw()
 
+    # the pitchItems list contains a tuple that represents all of the tick marks
+    # and text of the Pitch graducations.  The first item of the tuple is the angle
+    # and the second is the item reference.  We use this to make the tick marks
+    # disappear when they are a certain distance from the current pitch angle
+    def setPitchItems(self):
+        for each in self.pitchItems:
+            if abs(each[0] - self._pitchAngle) < self.visiblePitchAngle:
+                each[1].setOpacity(self.pitchOpacity)
+            else:
+                each[1].setOpacity(0)
+
     def redraw(self):
-        now = time.time()
-        if now - self.last_update_time < self.update_period:
-            return
-        self.last_update_time = now
         self.resetTransform()
-        if self.fdondb.value and self.fdtarget_widget is not None:
-            self.fdtarget_widget.update (self.fdpitchdb.value, self.fdrolldb.value)
         self.centerOn(self.scene.width() / 2,
                       self.scene.height() / 2 +
                       self._pitchAngle * self.pixelsPerDeg * - 1.0)
@@ -245,24 +263,33 @@ class AI(QGraphicsView):
         super(AI, self).paintEvent(event)
         w = self.width()
         h = self.height()
+        r = self.bankAngleRadius
         p = QPainter(self.viewport())
         p.setRenderHint(QPainter.Antialiasing)
+        # Put the static overlay image on the view
+        p.drawImage(self.rect(), self.overlay)
 
-        p.setPen(QColor(Qt.black))
-        p.setBrush(QColor(Qt.yellow))
-        p.drawRect(w / 4, h / 2 - 3, w / 6, 6)
-        p.drawRect(w - w / 4 - w / 6, h / 2 - 3, w / 6, 6)
-        p.drawRect(w / 2 - 3, h / 2 - 3, 8, 8)
+        pen = QPen(Qt.white)
+        pen.setWidth(1)
+        p.setPen(pen)
+        p.setBrush(QColor(Qt.white))
+        marks = QPen(Qt.white)
 
-        # Add non-moving Bank Angle Markers
-        marks = QPen(self.overlayColor)
-        marks.setWidth(3)
         p.translate(w / 2, h / 2)
+
+        # Slip / Skid ball
+        p.setPen(QColor(Qt.black))
+        p.setBrush(QColor(Qt.white))
+        p.drawEllipse(QPoint(self._latAccel * -120, -r + 32), 6, 6)
+
+        p.rotate(self._rollAngle * -1.0)
+        # Add moving Bank Angle Markers
+        marks.setWidth(3)
         p.setPen(marks)
         smallMarks = [10, 20, 45]
         largeMarks = [30, 60]
-        shortLine = QLine(0, - (h / 3), 0, - (h / 3 - 10))
-        longLine = QLine(0, - (h / 3 + 10), 0, - (h / 3 - 10))
+        shortLine = QLine(0, -r, 0, -(r-10))
+        longLine = QLine(0, -(r+10), 0, -(r-10))
         for angle in smallMarks:
             p.rotate(angle)
             p.drawLine(shortLine)
@@ -275,41 +302,24 @@ class AI(QGraphicsView):
             p.rotate(- 2 * angle)
             p.drawLine(longLine)
             p.rotate(angle)
-        if self.show_standard_turn and (not(self.fix_tas.fail)) and self.fix_tas.value > 30:
-            srpen = QPen(self.srTurnColor)
-            srpen.setWidth(3)
-            p.setPen(srpen)
-            #math.tan(_bank_angle_) = (rate_of_turn * tas_knots * 493/900) / 9.8(m/s)
-            # standard rate of turn = 3 degrees per second
-            rate_of_turn = math.radians(3)
-            tas_knots = self.fix_tas.value
-            #_bank_angle_ = math.atan((rate_of_turn * tas_knots * 493/900) / 9.8(m/s))
-            srbank = math.atan((rate_of_turn * tas_knots * 493/900) / 9.8)
-            srbank = math.degrees(srbank)
-            p.rotate(srbank)
-            p.drawLine(longLine)
-            p.rotate(- 2 * srbank)
-            p.drawLine(longLine)
-            p.rotate(srbank)
-
-        pen = QPen(self.overlayColor)
-        pen.setWidth(1)
-        p.setPen(pen)
-        p.setBrush(QColor(Qt.white))
-        triangle = QPolygon([QPoint(0 + 5, - (h / 3)),
-                             QPoint(0 - 5, - (h / 3)),
-                             QPoint(0, - (h / 3 - 10))])
+        triangle = QPolygon([QPoint(5, -r),
+                             QPoint(-5, -r),
+                             QPoint(0, -(r - 10))])
         p.drawPolygon(triangle)
-
-        p.rotate(self._rollAngle * -1.0)
-        triangle = QPolygon([QPoint(0 + 7, - (h / 3) + 25),
-                             QPoint(0 - 7, -(h / 3) + 25),
-                             QPoint(0, - (h / 3 - 10))])
-        p.drawPolygon(triangle)
-        pen = QPen(QColor(Qt.magenta))
-        pen.setWidth(3)
-        p.setPen(pen)
-        p.setBrush(QBrush())
+        # Draw standard rate turn markers
+        if self.drawBankMarkers:
+            a = math.degrees(math.atan(self._tas/364.0))
+            if a > self.bankAngleMaximum:
+                a = self.bankAngleMaximum
+            diamond = QPolygon([QPoint(0, -r),
+                                QPoint(-7, -r - 7),
+                                QPoint(-0, -r - 14),
+                                QPoint(7, -r - 7)])
+            p.setBrush(Qt.transparent)
+            p.rotate(a)
+            p.drawPolygon(diamond)
+            p.rotate(-2 * a)
+            p.drawPolygon(diamond)
 
     # We don't want this responding to keystrokes
     def keyPressEvent(self, event):
@@ -328,6 +338,17 @@ class AI(QGraphicsView):
         return self._rollAngle
 
     rollAngle = property(getRollAngle, setRollAngle)
+
+    def setLateralAcceleration(self, value):
+        if value != self._latAccel and self.isVisible():
+            self._latAccel = common.bounds(-0.3, 0.3, value)
+            self.update()
+
+    def setTrueAirspeed(self, value):
+        if value != self._tas and self.isVisible():
+            self._tas = value
+            self.update()
+
 
     def setAIFail(self, fail):
         log.debug("Set AI Fail")
@@ -374,6 +395,7 @@ class AI(QGraphicsView):
     def setPitchAngle(self, angle):
         if angle != self._pitchAngle and self.isVisible() and (not self._AIFail):
             self._pitchAngle = common.bounds(-90, 90, angle)
+            self.setPitchItems()
             self.redraw()
 
     def getPitchAngle(self):
